G21C-1735: Changed Mileage and SteeringAngle method signatures.

+++ PartnerLibrary/partnerlibrary/src/main/java/technology/cariad/partnerlibrary/MileageListener.java
@@ -1,5 +1,5 @@
 package technology.cariad.partnerlibrary;
 
 public interface MileageListener {
-    public abstract void onMileageValueChanged(int mileageValue);
+    public abstract void onMileageValueChanged(float mileageValue);
 }
