+++ PartnerLibrary/partnerlibrary/src/main/java/technology/cariad/partnerlibrary/SteeringAngleListener.java
@@ -1,5 +1,5 @@
 package technology.cariad.partnerlibrary;
 
 public interface SteeringAngleListener {
-    public abstract void onSteeringAngleChanged(int steeringAngle);
+    public abstract void onSteeringAngleChanged(float steeringAngle);
 }
